+++ FeatureExtractionNew/FeatureExtractionNew.ino/libmfcc.h
@@ -11,19 +11,19 @@
 #include <vector>
 #include "Cmplx.h"
 
-const double pi = 3.14159265358979323846264338327;
+const float pi = 3.14159265358979323846264338327;
 
 // Returns the specified (mth) MFCC
-double getCoefficient(std::vector<co::cd>& spectralData, unsigned int samplingRate, unsigned int NumFilters, unsigned int binSize, unsigned int m);
+float getCoefficient(std::vector<co::cd>& spectralData, unsigned int samplingRate, unsigned int NumFilters, unsigned int binSize, unsigned int m);
 
 // Compute the normalization factor (For internal computation only - not to be called directly)
-double NormalizationFactor(int NumFilters, int m);
+float NormalizationFactor(int NumFilters, int m);
 
 // Compute the filter parameter for the specified frequency and filter bands (For internal computation only - not the be called directly)
-double GetFilterParameter(unsigned int samplingRate, unsigned int binSize, unsigned int frequencyBand, unsigned int filterBand);
+float GetFilterParameter(unsigned int samplingRate, unsigned int binSize, unsigned int frequencyBand, unsigned int filterBand);
 
 // Compute the band-dependent magnitude factor for the given filter band (For internal computation only - not the be called directly)
-double GetMagnitudeFactor(unsigned int filterBand);
+float GetMagnitudeFactor(unsigned int filterBand);
 
 // Compute the center frequency (fc) of the specified filter band (l) (For internal computation only - not the be called directly)
-double GetCenterFrequency(unsigned int filterBand);
+float GetCenterFrequency(unsigned int filterBand);
